Add fingerprint function
The function helps to control schema version.

Fingerprint function is compatibla with any hash from Tarantool's digest
hashing library.

This function produces the same value as spache implementations on simple
cases, but may differ some times. (e.g. in cases with type references,
which copies type in this implementation)

The fingerprint is sustainable and the same for different representations
of the same schema.

Closes #30