Add forward_reference for an avro type
This change is not back-compatible and therefore is added with the
boolean option `forward_definition`.

The idea of the implementation is the following:
1. Allocate blank table for the type definition and its nullable
equivalent on the first reference.
2. Use this table for any references to the type.
3. Fill the table with type definition when the one is found.
4. At the end of schema creation:
4.1 Make sure that all blank types are filled with a type.
4.2 After copy is completed, update fields in nullable and non nullable
type versions (only one of them is filled with data during compilation
stage)

Problems which influenced the implementation:
1. Exporting table mechanism uses node tables as keys to keep track of
types that are already exported. That meant that all references to a
type should be represented with the very same table in AST (not a copy).

Side effects:
1. An exported schema is always represented in canonical form. That
means that In case of deferred type reference, exported schema contains
no more deferred definitions.

Closes #48