Fix fingerprint for type references
The fingerprint calculation is performed over json canonical form of the
schema, which means that the whole schema should have only one type
definition for each type, the following references should reference the
type by name.
The easiest way to do this was to reuse export function, which prepares
the canonical form of the schema for the output.

Closes #52