Use type name for records to prevent cycles
Using tables as keys of the dictionary to prevent cycles lead to an
error in case when there are several tables for the same type.  This
situation is possible by now due to nullable fields and following
commits may introduce even more such cases.

Cycles for other types are still checked using tables as a keys.
However, there should not be cycles which do not include records.