Give variables the same names in ir and in lua
It is very important for debug purposes for variables to have the same
names in the ir and resulting lua representation.

This change leads to creating more variables in the resulting lua code.
The maximal number of extra variables equals to number of service
fields.

Allocating extra variables should not affect performance.